requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
fastapi>=0.115.5
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
